poc-python-firestore/src/functions/requirements.txt
requests==2.22.0
google-auth==1.6.3
urllib3==1.25.3